Allow a blacklist listing filters
We want to preserve active generation of filters, while
opting out of certain filters, for example associations.

Sometimes associations can be prohibitively large, for
example when there are 100,000 users and several models
which belong_to :user.  That filter would have too many 
items for the list to display correctly, and was causing 
serious performance issues.