Added #normalized to help in simple matching in auth adapters.
The `#authorized?` method's subject can be set to both instances as well
as classes of objects. This can make it much difficult to create simple
case statements for authorization since you have to handle both the
class level match and the instance level match.

For example:

    class MyAuthAdapter < ActiveAdmin::AuthorizationAdapter

      def authorized?(action, subject = nil)
        case subject
        when Post
          true
        when Class
          if subject == Post
            true
          end
        end
      end

    end

To handle this, the normalized method takes care of returning a lambda
which implements `===` to be matched in a case statement.

The above now becomes:

    class MyAuthAdapter < ActiveAdmin::AuthorizationAdapter

      def authorized?(action, subject = nil)
        case subject
        when normalized(Post)
          true
        end
      end

    end