ActiveAdmin::Resource loads resources for breadcrumbs
Closes #2266

Add `ActiveAdmin::Resource#find_resource`. This is now a centralized place to
load a resource, with correct handling of primary key, decoration, etc.

Then, use this new way of loading resources in the Breadcrumbs logic.
This was the original impetus for for defining
`ActiveAdmin::Resource#find_resource`. Previously, the `BreadcrumbsHelper`
did a one-off load of the resource, which has historically been a source of
bugs and/or inconsistencies.

Please note that this is duplication with InheritedResources. Normally, the
resource is loaded by InheritedResources, but there are situations in which
ActiveAdmin needs to load resources as well (ie Breadcrumbs). It was not
reasonable to use InheritedResources to load the resource (see the discussion
at https://github.com/gregbell/active_admin/pull/2315).