get search method validation "working" with Ransack
With MetaSearch, if you asked if the search object responded to a
method, that meant that it was a valid search method. Ransack doesn't
behave that way, so all we can really do is guess.

Ransack instead returns true if the method is valid **OR** if the
method equals the name of one of the DB attributes, which in my opinion
is completely useless.