Use `#has_attribute?` as suggested by deprecation warning
The `#is_boolean?` implementation calls `#type_for_attribute` or
`#column_for_attribute` depending on the Rails version. However,
`#type_for_attribute` returns an instance of `ActiveRecord::Type::Value`
with a null `#type`, which makes the `#status_tag` tests fail.

Trying to change it back to `#column_for_attribute` results in the
following deprecation warning:

    DEPRECATION WARNING: `#column_for_attribute` will return a null
    object for non-existent columns in Rails 5. Use `#has_attribute?` if
    you need to check for an attribute's existence.

This commit follows this suggestion and uses `#has_attribute?` before
accessing the column information with `#column_for_attribute`, which
should be compatible with all major Rails versions supported.

The deprecation warning will still be there, but the null check is in
place as well.