Fixes #3583: Associates comments with STI base classes
The Rails convention of a polymorphic belongs_to is to associate with the base class
not the subclass.  The makes the association more durable, in the case that a model instance
changes its type (switches subclasses).  In addition, all comments through 0.6.x were
associated with the resource base class.  The change to associate with the subclass has
orphaned those Comments.  This change fixes those problems.