Fix failing scenarios in jruby
I'm not sure why but somehow some assertions don't seem to wait for the
content to appear in jruby. Migrating this tests to use a javascript
driver fixes it.

Furthermore, I think this is a good change (except for performance)
because this tests are actually supposed test js behaviour. The only
reason they were passing with a non-js driver is that they manually
simulate the js behaviour. See the "When I submit the batch action form"
step for an example. It's not testing the real behavior of the user, so
if any regressions are introduced where that stuff actually happens
(batch_actions.js.coffee), they won't be catched by the test.

Regarding the actual migration of the tests, it includes a couple of
changes to avoid using :contains or :first that are not valid css
selectors and are only working "by chance" (see [1], last comment) and a
rewrite of the previously mentioned step at a higher level to actually
trigger the js that is being tested.

[1]: https://github.com/teampoltergeist/poltergeist/issues/323