Stop testing against Rails 4.1
We’re getting this error:

```
undefined method `to_h' for {}:ActionController::Parameters
```

The odd thing is Rails 4.1 on Ruby 2 and JRuby are passing just fine.
It’s only Rails 4.1 on Ruby 1.9 that’s failing.

I don’t think it’s important to run the test suite on Rails 4.1, and
we’ll be dropping support for it soon anyway, so this commit removes it
from Travis.