Make the has_many allow_destroy option smart
When I generate nested has_many forms, I might whant to set the `allow_destroy` parameter
depending on the child record.

For example: I want to not allow delete comments, that was created by an admin,
or I want to not allow delete articles that are published.

I added the ability to specify Symbol and Proc as `allow_destroy` parameter.

If a Symbol is specified, the child object's method will be called.

    # this will create 'delete' checkbox only for articles which #draft? method returns true
    f.has_many :articles, allow_destroy: :draft? do |a|

If a Proc is specified, the given blok will be calculated  for the child object.
The object is passed as the parameter of the block.

    f.has_many :comments, allow_destroy:  -> (comment) { !comment.author.admin? } do |c|