Don't gitignore bin
`bin` is the default "bindir" folder in the gemspec. Executables
provided by a gem should live in this folder. In this case, the gem
provides no executables, but in my case (don't know why) there was a
`bin/console` script in this folder.

I figure if we gitignore this folder, there's a chance that we might
unintentionally ship scripts in `bin/` with the gem.