Move appraisal gem into development group
And make sure it's not automatically required.

We are experiencing segfaults during CI builds with ruby 2.4.0 inside
appraisal code. Appraisal is, however, not necessary for those builds
since Travis simply sets the `BUNDLE_GEMFILE` environment variable. So I
made sure Appraisal is not actually used there.

We don't need to worry about this in development since when Appraisal is
actually used (gemfile management, running tests against a specific
Rails version without using BUNDLE_GEMFILE directly), it'll run
directly (through appraisal binstub).

We should report the crash to Appraisal, though, so they can investigate
and forward it upstream if it happens to be a Ruby bug. It's happing in an
`instance_eval`, so a wild guess could be
https://bugs.ruby-lang.org/issues/13090.