Use "starts/ends with" placeholders for date range fields
This is all around a better choice as its descriptive what each field
does in place not having a label like other filter fields do. We need
to apply box-sizing: border-box on the input so they stay evenly side
by side because the width we give it is the width we want it to have.
Without that rule, the padding is in addition to the width (meaning 5px
horizontal padding and 100px width, then the rendered width is 110px).