Add ActiveAdmin class with helpers
Going with something simple like this so we can pass this around safely
since we don’t have a supported way to do exports here using
sprockets-es6. This consolidates some similar code reused in a few
places dealing with Turbolinks and query strings. For the serializing
and deserializing of query strings I stuck with the array of objects
approach as that is better one to support so we don’t have to combine
objects if name matches.