Workaround issue with base image
Base ruby docker images where setting the `BUNDLER_VERSION` environment
variable. This environment variable is used in bundler 2 to override
"lock file bundler version guessing". Setting this environment variable
has been removed from the official docker images, so that it doesn't
interfere with this feature, but it has not yet been released. So for
now we override the `BUNDLER_VERSION` environment variable with the
value we want, so that bundler can properly run.