Switch sass dependency to sassc-rails (#5504)
With sprockets v4, we might be able to remove this dependency and depend
only on `sassc`. But at the moment, sprockets v3 autoloads `sass`, which
we no longer depend on so we need to pull `sassc-rails` to do the proper
modifications to sprockets to not do that and load `sassc` instead.