Separate status_tag `nil` case for optional locale customization
When using the `status_tag` component with boolean fields, both `nil` and `false` were treated the same and had the same output but without a way to modify that other than overriding the component.

These 2 conditions are now split, so for the `nil` case it returns "Unset" but the locale will keep a display value of "No" for that locale key. That **will maintain backwards compatibility** so when `status_tag` is called with `nil` or `false` it will continue to display the same "No" label but provide the option to customize the label for the `nil` case only, leaving the `false` one as is.

Further, we need to ensure the same CSS class for presentation due to the change, so the class list output now includes "no" and "unset".